refactor(apex-lsp-compliant-services): extract rename subsystem into worker/rename.ts - W-24145209
Move the local / field / method rename + prepareRename resolution (~1.8k
lines: resolveLocalRename, resolveFieldRename, resolveMethodRename, the three
resolvePrepareRename* entries, their context/verify helpers, provenance guard
and rename result types) out of worker.platform.shared.ts into
src/worker/rename.ts.

None of these functions were part of the module's public surface; the
composition root imports the six dispatch entry points and wires them into the
DispatchRename / DispatchPrepareRename handlers, exactly as before.

rename.ts leans on the shared cursor/occurrence resolution substrate
(resolveCursorSymbol, recompileCursorFileAtFullDetail, targetSymbolForCursor,
declarationLocationForCursor, loadReferencedTypesForFile, exactCursorReference,
field/method declaration-range helpers), which still lives in the shared
module and is imported back. Every reference is inside a function body run at
request time, never at module init, so this back-edge to the composition root
is evaluation-safe (and import/no-cycle is not enabled). resolveCursorSymbol,
methodDeclarationRangeFromParse and exactCursorReference gained `export` so the
new module can consume them.

Behavior is unchanged; the full unit + integration matrix (incl. the
renameField / renameMethod worker-topology tests) stays green.

W-24145209
